Expand a dataset to include other levels of aggregation

Uses the structural specification given in .spec to aggregate a time series. A grouped structure is specified using grp1 * grp2, and a nested structure is specified via parent / child. Aggregating the key structure is commonly used with forecast reconciliation to produce coherent forecasts over some hierarchy.

aggregate_key(.data, .spec, ...)

Arguments

.data
A tsibble.

.spec
The specification of aggregation structure.

...
Name-value pairs of summary functions. The name will be the name of the variable in the result. The value should be an expression that returns a single value like min(x), n(), or sum(is.na(y)).

The arguments in ... are automatically quoted and evaluated in the context of the data frame. They support unquoting and splicing. See vignette("programming") for an introduction to these concepts.

Details

This function is experimental, and is subject to change in the future.

The way in which the measured variables are aggregated is specified in a similar way to how [dplyr::summarise()] is used.

See also

Examples

library(tsibble)
tourism %>%
  aggregate_key(Purpose * (State / Region), Trips = sum(Trips))

#> # A tsibble: 34,000 x 5 [1Q]
#> # Key:       Purpose, State, Region [425]
#>    Purpose      State        Region       Quarter  Trips
#>    <chr>        <chr>        <chr>          <qtr>  <dbl>
#>  1 <aggregated> <aggregated> <aggregated> 1998 Q1 23182.
#>  2 <aggregated> <aggregated> <aggregated> 1998 Q2 20323.
#>  3 <aggregated> <aggregated> <aggregated> 1998 Q3 19827.
#>  4 <aggregated> <aggregated> <aggregated> 1998 Q4 20830.
#>  5 <aggregated> <aggregated> <aggregated> 1999 Q1 22087.
#>  6 <aggregated> <aggregated> <aggregated> 1999 Q2 21458.
#>  7 <aggregated> <aggregated> <aggregated> 1999 Q3 19914.
#>  8 <aggregated> <aggregated> <aggregated> 1999 Q4 20028.
#>  9 <aggregated> <aggregated> <aggregated> 2000 Q1 22339.
#> 10 <aggregated> <aggregated> <aggregated> 2000 Q2 19941.
#> # … with 33,990 more rows