Decomposition modelling

This function allows you to specify a decomposition combination model using any additive decomposition. It works by first decomposing the data using the decomposition method provided to dcmp_fn with the given formula. Secondary models are used to fit each of the components from the resulting decomposition. These models are specified after the decomposition formula. All non-seasonal decomposition components must be specified, and any unspecified seasonal components will be forecasted using fable::SNAIVE(). These component models will be combined according to the decomposition method, giving a combination model for the response of the decomposition.

decomposition_model(dcmp_fn, formula, ..., dcmp_args = list())

Arguments

dcmp_fn
The decomposition function (such as feasts::STL())

formula
The formula used to describe the decomposition

...
Model definitions used to model the components (such as fable::ETS())

dcmp_args
Arguments to be passed to the decomposition function (.dcmp_fn)

See also

Examples

library(fable)

#> Error: package ‘fablelite’ required by ‘fable’ could not be found

library(tsibble)
library(dplyr)
library(feasts)

#> Error: package ‘fablelite’ required by ‘feasts’ could not be found

vic_food <- tsibbledata::aus_retail %>%
  filter(State == "Victoria", Industry == "Food retailing")

# Identify an appropriate decomposition
vic_food %>%
  STL(log(Turnover) ~ season(window = Inf)) %>%
  autoplot()

#> Error in STL(., log(Turnover) ~ season(window = Inf)): could not find function "STL"

# Use an ARIMA model to seasonally adjusted data, and SNAIVE to season_year
# Any model can be used, and seasonal components will default to use SNAIVE.
my_dcmp_spec <- decomposition_model(
  STL, log(Turnover) ~ season(window = Inf),
  ETS(season_adjust), SNAIVE(season_year)
)

#> Error in ETS(season_adjust): could not find function "ETS"

vic_food %>%
  model(my_dcmp_spec) %>%
  forecast(h="5 years") %>%
  autoplot(vic_food)

#> Error in dots_list(...): object 'my_dcmp_spec' not found