Produce forecasts

The forecast function allows you to produce future predictions of a time series from fitted models. If the response variable has been transformed in the model formula, the transformation will be automatically back-transformed (and bias adjusted if bias_adjust is TRUE). More details about transformations in the fable framework can be found in vignette("transformations", package = "fable").

forecast(object, ...)

# S3 method for mdl_df
forecast(object, new_data = NULL, h = NULL,
  bias_adjust = TRUE, ...)

Arguments

object
The time series model used to produce the forecasts

...
Additional arguments for forecast model methods.

new_data
A tsibble containing future information used to forecast.

h
The forecast horison (can be used instead of new_data for regular time series with no exogenous regressors).

bias_adjust
Use adjusted back-transformed mean for transformations. Refer to vignette("transformations", package = "fable") for more details.

Details

The forecasts returned contain both point forecasts and their distribution. A specific forecast interval can be extracted from the distribution using the hilo() function, and multiple intervals can be obtained using report(). These intervals are stored in a single column using the hilo class, to extract the numerical upper and lower bounds you can use tidyr::unnest().

Examples

library(fable)

#> Error: package ‘fablelite’ required by ‘fable’ could not be found

library(tsibble)
library(tsibbledata)
library(dplyr)
library(tidyr)

# Forecasting with an ETS(M,Ad,A) model to Australian beer production
beer_fc <- aus_production %>%
  model(ets = ETS(log(Beer) ~ error("M") + trend("Ad") + season("A"))) %>%
  forecast(h = "3 years")

#> Error in ETS(log(Beer) ~ error("M") + trend("Ad") + season("A")): could not find function "ETS"

# Compute 80% and 95% forecast intervals
beer_fc %>%
  hilo(level = c(80, 95))

#> Error in eval(lhs, parent, parent): object 'beer_fc' not found

beer_fc %>%
  autoplot(aus_production)

#> Error in eval(lhs, parent, parent): object 'beer_fc' not found

# Forecasting with a seasonal naive and ETS(A,A,A) model to the monthly 
# "Food retailing" turnover for each Australian state/territory.
library(dplyr)
aus_retail %>%
  filter(Industry == "Food retailing") %>%
  model(
    snaive = SNAIVE(Turnover),
    ets = ETS(log(Turnover) ~ error("A") + trend("A") + season("A")),
  ) %>%
  forecast(h = "2 years 6 months") %>%
  autoplot(filter(aus_retail, Month >= yearmonth("2000 Jan")), level = 90)

#> Error in SNAIVE(Turnover): could not find function "SNAIVE"

# Forecast GDP with a dynamic regression model on log(GDP) using population and
# an automatically chosen ARIMA error structure. Assume that population is fixed
# in the future.
aus_economy <- global_economy %>%
  filter(Country == "Australia")
fit <- aus_economy %>%
  model(lm = ARIMA(log(GDP) ~ Population))

#> Error in ARIMA(log(GDP) ~ Population): could not find function "ARIMA"

future_aus <- new_data(aus_economy, n = 10) %>%
  mutate(Population = last(aus_economy$Population))

fit %>%
  forecast(new_data = future_aus) %>%
  autoplot(aus_economy)

#> Error in eval(lhs, parent, parent): object 'fit' not found